+++ src/screens/home/news/HomeNewsRowViewModel.ts
