+++ src/screens/home/news/HomeNewsRowViewModelMapper.ts
@@ -1,7 +1,7 @@
-import { News } from '../../core/entities/News'
+import { News } from '../../../core/entities/News'
 import { format } from 'date-fns'
 import { HomeNewsRowViewModel } from './HomeNewsRowViewModel'
-import i18n from '../../utils/i18n'
+import i18n from '../../../utils/i18n'
 
 export const HomeNewsRowViewModelMapper = {
   map: (news: News): HomeNewsRowViewModel => {
