+++ src/screens/home/quickPoll/HomeQuickPollRowAnswerViewModel.ts
