+++ src/screens/home/quickPoll/HomeQuickPollRowContainer.tsx
@@ -1,10 +1,10 @@
 import React, { FunctionComponent } from 'react'
 import { StyleSheet, Text, View } from 'react-native'
-import { Colors, Spacing, Typography } from '../../styles'
-import CardView from '../shared/CardView'
+import { Colors, Spacing, Typography } from '../../../styles'
+import CardView from '../../shared/CardView'
 import HomeQuickPollChoicesView from './HomeQuickPollChoicesView'
 import HomeQuickPollResultView from './HomeQuickPollResultView'
-import { HomeQuickPollRowContainerViewModel } from './HomeRowViewModel'
+import { HomeQuickPollRowContainerViewModel } from '../HomeRowViewModel'
 
 type Props = Readonly<{
   viewModel: HomeQuickPollRowContainerViewModel
